requirements.txt
fastapi[standard]
starlette
SQLAlchemy
alembic
pytz
python-jose
passlib
pydantic
symspellpy
python-docx
python-dotenv
pybloom_live
pymysql
passlib[bcrypt]
passlib[argon2]
cryptography